1.1.4 -- peer body physics was never actually trimmed
The work in 1.1.1 that was supposed to make other players' body physics cheap
has been dying on the second body it touched, at every mount, since the day it
shipped. It applied nothing.

The cause was a parameter one byte wide where the engine reads four.
FBodyInstance::SetResponseToChannel forwards plain C++ enums, and
FCollisionResponse::SetResponse reads all 32 bits of them -- it sign-extends the
channel with movsxd and indexes a global table with the result. The declaration
here said uint8_t, which only obliges the compiler to write the low byte.
Channel 0 survived by luck, because 0 is materialised as an instruction that
clears the whole register; channel 1 is a byte store that leaves the upper 24
bits as they were, so the table index went wild and read unmapped memory. The
failure advanced exactly one body per attempt -- the first call had already
written its channel, so that body was skipped next pass -- and it was that
staircase in the log that gave it away.

With it fixed, a peer riding their board goes from 17 of 21 physics bodies
simulated on the watching machine to 13, all 21 stop colliding against the
level, and their solver iterations halve. None of that was happening before.

Two smaller things from the same investigation:

A body whose bone cannot be identified yet -- which happens when the first pass
of a mount lands before the skeleton is readable -- was cached as "not a leg"
for the rest of the ride. It is now retried; a body that cannot be identified is
still never touched.

The fault handler said only "body trim FAULTED", which is why this went a whole
release unnoticed. It now names the body, the step, the exception code and the
exe-relative address.

Also measured and ruled out: the physical-animation rig rebuild on a mount is
0.06 ms, and the per-tick target push is about 6 us per peer per frame. Neither
is the mount hitch. On a riding proxy 20 of 21 bodies carry real blend weight,
the same as the owner's, so there is no discarded simulation to skip. The cost
is the solve itself, and simulating less of it is the only lever.

+++ src/game/game_syms.h
@@ -261,7 +261,14 @@ using BoardRefreshVisFn = void (*)(void* boardActor);
 // is safe either way, since a Microsoft x64 caller that hands over an argument the callee does not
 // read has simply set a register nobody looks at.
 using BodySetSimulateFn  = void (*)(void* bodyInstance, bool simulate, bool maintainBlend, bool preserveAttach);
-using BodySetResponseFn  = void (*)(void* bodyInstance, uint8_t channel, uint8_t response);
+// ECollisionChannel and ECollisionResponse are plain C++ enums, so they are THIRTY-TWO BITS on the
+// wire, and FCollisionResponse::SetResponse reads them as such -- `movsxd rbp, edx` sign-extends the
+// whole register and indexes a global table with it, and `cmp eax, r8d` compares all 32 bits. These
+// were declared uint8_t, which only obliges the compiler to write the low byte: channel 0 came out
+// clean because it is materialised as `xor edx,edx`, but channel 1 as `mov dl,1` left whatever was
+// already in the upper 24 bits, and the table index went wild. That is the access violation the peer
+// body trim died on at every mount since it shipped. Never narrow a parameter the callee reads wide.
+using BodySetResponseFn  = void (*)(void* bodyInstance, int32_t channel, int32_t response);
 using BodyUpdateFilterFn = void (*)(void* bodyInstance);
 // FPhysicsCommand_PhysX::ExecuteWrite(const FPhysicsActorHandle&, TFunctionRef<void(const handle&)>).
 // The TFunctionRef is two pointers, thunk first, and is called as thunk(callable, handleRef).
